IMP product_pack format
