+++ product_pack/__init__.py
@@ -3,4 +3,6 @@
 # For copyright and license notices, see __openerp__.py file in root directory
 ##############################################################################
 from . import pack
-from . import sale
+from . import sale_order_line_pack_line
+from . import sale_order_line
+from . import sale_order
