+++ product_pack/__openerp__.py
@@ -27,10 +27,14 @@
     'description': """
 Product Pack
 ============
-# TODO agregar en configuracion si se quiere usar los sale order packs (seria para el group group_pack) y ver que se haga visible la vista form
+# TODO agregar en configuracion si se quiere usar los sale order packs
+(seria para el group group_pack) y ver que se haga visible la vista form
 # TODO implementar totalice en price get
-# TODO agregar constraint de no pack dentro de pack
+# TODO ver de permitir packs dentro de packs (robar de v7)
 # TODO calcular correctamente pack virtual available para negativos
+# TODO chequear si en todos los casos un copy de sale order anda bien o hacemos
+como en v7 que el copy lo bloqueamos
+
     """,
     'author':  'NaN·tic, ADHOC',
     'images': [
